requirements.txt
APScheduler==3.6.3
backports-abc==0.5
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
ipaddress==1.0.23
pycparser==2.20
python-telegram-bot==13.1
pytz==2021.1
singledispatch==3.4.0.3
six==1.15.0
tornado==5.1.1
tzlocal==2.1
